+++ doc/changelog.rst
@@ -1,6 +1,17 @@
 Changelog
 ======
 
+
+v0.5.0 - 17.05.2026
+------
+* Updated the object detection function to support detecting objects of all colors.
+* Added helper function that reads the object's RGB and HSV color values from a selected pixel of the camera image.
+* Added program skeleton intended for testing the Kalman filter.
+
+v0.5 - Kalman filter
+======
+
+
 v0.4.9 - 13.05.2026
 ------
 * Added test for the robot's movement while maintaining constant force.